+++ src/sha1.c
@@ -207,8 +207,8 @@ void VDB_SHA1_Update(VDB_SHA1_CTX* context, const uint8_t* data, const size_t le
 #endif
 
     j = (context->count[0] >> 3) & 63;
-    if ((context->count[0] += len << 3) < (len << 3)) context->count[1]++;
-    context->count[1] += (len >> 29);
+    if ((context->count[0] += (uint32_t)len << 3) < (len << 3)) context->count[1]++;
+    context->count[1] += ((uint32_t)len >> 29);
     if ((j + len) > 63) {
         memcpy(&context->buffer[j], data, (i = 64-j));
         SHA1_Transform(context->state, context->buffer);
